modularize dnn simplification; add various typing
