+++ dnnv/nn/operations/__init__.py
@@ -5,5 +5,4 @@ from .nn import *
 from .patterns import *
 from .tensor import *
 
-Activation: OperationPattern = Relu | Sigmoid
-MatrixMultiplication: OperationPattern = Gemm | MatMul
+Activation: OperationPattern = Relu | Sigmoid | Tanh
