+++ dnnv/verifiers/eran/layers.py
@@ -6,7 +6,6 @@ from typing import Optional, Type, Union
 from dnnv import logging
 from dnnv.nn.layers import Layer, Convolutional
 from dnnv.nn.operations import (
-    Activation,
     Add,
     BatchNormalization,
     Conv,
